add a route /rooms that reads directly from the DB and serializes it to JSON. Also maintain a version of the database on disk so that we can know when to wipe and rebuilt the tables

+++ marshaller.go
@@ -2,21 +2,20 @@ package main
 
 import (
 	"database/sql"
-	"fmt"
 	_ "github.com/mattn/go-sqlite3"
 	"log"
 	"os"
+	"io/ioutil"
 )
 
 type Room struct {
-	id int
-    name string
+	// capitalized variables so they are exported https://golang.org/ref/spec#Exported_identifiers
+	Id int `json:"id"`
+    Name string `json:"name"`
 }
 
 
 type Song struct {
-    name string // do we need this?
-    videoLink string // Can we retrieve all the information from this link? Do we need the other params
     videoData string // do we need this?
 }
 
@@ -27,60 +26,75 @@ room->DeleteSong(videoLink)
 room->AddSong(videoLink)
 */
 
+// keep a versioning scheme for the db so we can know when to recreate the sqlite file
+const dbVersion string = "1";
 
+// TODO (ajafri): we perform a file read on each one of these calls so use it sparingly or change the pattern
 func createNewDB() (*sql.DB, error) {
-	// TODO (ajafri): Should we be clearing the database on startup. Probably not.
-	os.Remove("./rooms.db")
+	var version string = "dbVersion"
+
+	b, err := ioutil.ReadFile("bridgdDBVersion.txt")
+    if err != nil {
+        b = []byte("") // simulate a version mismatch if reading the file throws an error
+    }
+
+    version = string(b)
+
+    if version != dbVersion {
+    	log.Println	("Version difference detected - recreating database")
+
+	    err = ioutil.WriteFile("bridgdDBVersion.txt", []byte(dbVersion), 0644)
+	    if err != nil {
+	        panic(err)
+	    }
+
+		os.Remove("./rooms.db") // clear the db if it is not versioned to the current version
+		
+		// let's create a new db and instantiate the structure 
+		db, err := sql.Open("sqlite3", "./rooms.db") 
+    	if err != nil {
+    		return db,err
+    	}
+
+    	var query string = "create table rooms (id integer not null primary key, name text);"
+		_, err = db.Exec(query)
+		if err != nil {
+			log.Printf("%q: %s\n", err, query)
+			return db, err
+		}
+
+		query = "insert into rooms(id, name) values(1, \"no yolo zone\")"
+		_, err = db.Exec(query)
+		if err != nil {
+			log.Printf("%q: %s\n", err, query)
+			return db, err
+		}
+
+    } 
+		
 	return sql.Open("sqlite3", "./rooms.db")
 }
 
-
-// TODO (ajafri): This is just an example of how we can read/write to the db. Take it out. 
-func writeToDB() {
+func getRooms() ([]Room) {
 	db, err := createNewDB()
 	if err != nil {
 		log.Fatal(err)
 	}
 	defer db.Close()
 
-	sqlStmt := `
-	create table rooms (id integer not null primary key, name text);
-	delete from rooms;
-	`
-	_, err = db.Exec(sqlStmt)
-	if err != nil {
-		log.Printf("%q: %s\n", err, sqlStmt)
-		return
-	}
-
-	tx, err := db.Begin()
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	stmt, err := tx.Prepare("insert into rooms(id, name) values(?, ?)")
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer stmt.Close()
-
-	// underscore here means discard the resultant. A variable you can write to but not read from. It will get compiled out.
-	_, err = stmt.Exec(1, fmt.Sprintf("This is bridgd room %d!!", 1))
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	tx.Commit()
-
 	rows, err := db.Query("select id, name from rooms")
 	if err != nil {
 		log.Fatal(err)
 	}
+
+	var rooms []Room
+
 	defer rows.Close()
 	for rows.Next() {
-		var id int
-		var name string
-		rows.Scan(&id, &name)
-		fmt.Println(id, name)
+		var room Room
+		rows.Scan(&room.Id, &room.Name)
+		rooms = append(rooms, room)
 	}
+
+	return rooms
 }
